Add bin dir

+++ link.sh
@@ -6,5 +6,7 @@ cd link
 
 for dotfile in *
 do
-  ln -sf ${PWD}/${dotfile} ${HOME}/.${dotfile}
+  if [[ -e ${HOME}/.${dotfile} ]]; then
+    ln -sf ${PWD}/${dotfile} ${HOME}/.${dotfile}
+  fi
 done
